src/requirement/main/it_ticket/pyproject.toml
[tool.poetry]
name = "it-ticket"
version = "0.1.0"
description = "IT_Ticket bot for PSCP"
authors = ["Kawaiiow <poohsn1357@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
